Move SSv3Alpha class & interface
